Removed DataQualityResult as it wasn't needed
