+++ waimak-deequ/src/main/scala/com/coxautodata/waimak/dataflow/spark/dataquality/deequ/DeequCheck.scala
@@ -1,6 +1,8 @@
 package com.coxautodata.waimak.dataflow.spark.dataquality.deequ
 
-import com.amazon.deequ.{VerificationRunBuilder, VerificationSuite}
+import com.amazon.deequ.checks.CheckStatus
+import com.amazon.deequ.constraints.{ConstraintResult, ConstraintStatus}
+import com.amazon.deequ.{VerificationResult, VerificationRunBuilder, VerificationSuite}
 import com.coxautodata.waimak.dataflow.spark.dataquality._
 import org.apache.spark.sql.Dataset
 
@@ -9,10 +11,42 @@ case class DeequCheck(checks: VerificationRunBuilder => VerificationRunBuilder) 
 
   override def ++(other: DeequCheck): DeequCheck = DeequCheck(checks andThen other.checks)
 
-  override def getResult(data: Dataset[_]): DataQualityResult = {
-    DeequResult(checks(VerificationSuite()
+  override def getAlerts(label: String, data: Dataset[_]): Seq[DataQualityAlert] = {
+    val verificationResult = getResult(data)
+    verificationResult.status match {
+      case CheckStatus.Success => Nil
+      case _ => verificationResult
+        .checkResults.values
+        .flatMap(
+          result => result.constraintResults
+            .filter(_.status != ConstraintStatus.Success)
+            .map(constraintResultToAlert(label, _, getAlertImportance(result.status)))
+        )
+        .toSeq
+    }
+  }
+
+  def getResult(data: Dataset[_]): VerificationResult = {
+    checks(VerificationSuite()
       .onData(data.toDF))
-      .run())
+      .run()
+  }
+
+
+  def constraintResultToAlert(label: String, constraintResult: ConstraintResult, alertImportance: AlertImportance): DataQualityAlert = {
+    val message =
+      s"""${alertImportance.description} alert for label $label
+         | ${constraintResult.constraint} : ${constraintResult.message.getOrElse("")}
+       """.stripMargin
+    DataQualityAlert(message, alertImportance)
+  }
+
+  def getAlertImportance(checkStatus: CheckStatus.Value): AlertImportance = {
+    checkStatus match {
+      case CheckStatus.Success => Good
+      case CheckStatus.Warning => Warning
+      case CheckStatus.Error => Critical
+    }
   }
 }
